Remove all `versioneer` related artifacts
`pyproject.toml` is configured to use `setuptools-scm` to generate
the version file, not `versioneer`. However, we still had an old
`versioneer` file at `abl/vpath/base/_version.py` which was
causing `git` to be invoked in release builds of Live (since we
would unconditionally try to get the version from `git` instead
of using the static version in release builds).